authentication done

+++ src/main/java/com/yash/journalApp/service/JournalEntryService.java
@@ -27,7 +27,7 @@ public class JournalEntryService {
             journalEntity.setDate(LocalDateTime.now());
             JournalEntry saved = journalEntryRepository.save(journalEntity);
             user.getJournalEntries().add(saved);
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch(Exception e) {
             System.out.println(e);
         }
@@ -47,8 +47,10 @@ public class JournalEntryService {
 
     public void deleteById(ObjectId id, String userName) {
         User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        boolean removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if(removed) {
+            userService.saveUser(user);
+            journalEntryRepository.deleteById(id);
+        }
     }
 }
